[break-them-up] Too big to fail is too big to exist
