+++ cmd/ddev/cmd/start.go
@@ -5,6 +5,7 @@ import (
 	"log"
 	"os"
 
+	"github.com/drud/ddev/pkg/dockerutil"
 	"github.com/drud/ddev/pkg/util"
 	"github.com/spf13/cobra"
 )
@@ -32,9 +33,9 @@ var StartCmd = &cobra.Command{
 			os.Exit(0)
 		}
 
-		client := util.GetDockerClient()
+		client := dockerutil.GetDockerClient()
 
-		err := util.EnsureNetwork(client, netName)
+		err := dockerutil.EnsureNetwork(client, netName)
 		if err != nil {
 			log.Fatalf("Unable to create/ensure docker network %s, error: %v", netName, err)
 		}
