+++ pkg/archive/archive.go
@@ -1,4 +1,4 @@
-package util
+package archive
 
 import (
 	"archive/tar"
@@ -6,10 +6,11 @@ import (
 	"compress/gzip"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/drud/ddev/pkg/util"
 )
 
 // Ungzip accepts a gzipped file and uncompresses it to the provided destination path.
@@ -72,7 +73,7 @@ func Untar(source string, dest string, extractionDir string) error {
 		return err
 	}
 
-	defer CheckClose(f)
+	defer util.CheckClose(f)
 
 	if strings.HasSuffix(source, "gz") {
 		gf, err := gzip.NewReader(f)
@@ -80,12 +81,12 @@ func Untar(source string, dest string, extractionDir string) error {
 			return err
 		}
 
-		defer CheckClose(gf)
+		defer util.CheckClose(gf)
 
 		tf = tar.NewReader(gf)
 	} else {
 		tf = tar.NewReader(f)
-		defer CheckClose(f)
+		defer util.CheckClose(f)
 	}
 
 	for {
@@ -152,7 +153,7 @@ func Unzip(source string, dest string, extractionDir string) error {
 	if err != nil {
 		return fmt.Errorf("Failed to open zipfile %s, err:%v", source, err)
 	}
-	defer CheckClose(zf)
+	defer util.CheckClose(zf)
 
 	for _, file := range zf.File {
 		// If we have an extractionDir and this doesn't match, skip it.
@@ -195,109 +196,4 @@ func Unzip(source string, dest string, extractionDir string) error {
 	}
 
 	return nil
-}
-
-// CopyFile copies the contents of the file named src to the file named
-// by dst. The file will be created if it does not already exist. If the
-// destination file exists, all it's contents will be replaced by the contents
-// of the source file. The file mode will be copied from the source and
-// the copied data is synced/flushed to stable storage. Credit @m4ng0squ4sh https://gist.github.com/m4ng0squ4sh/92462b38df26839a3ca324697c8cba04
-func CopyFile(src string, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer CheckClose(in)
-	out, err := os.Create(dst)
-	if err != nil {
-		return fmt.Errorf("Failed to create file %v, err: %v", src, err)
-	}
-	defer CheckClose(out)
-	_, err = io.Copy(out, in)
-	if err != nil {
-		return fmt.Errorf("Failed to copy file from %v to %v err: %v", src, dst, err)
-	}
-
-	err = out.Sync()
-	if err != nil {
-		return err
-	}
-
-	si, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-	err = os.Chmod(dst, si.Mode())
-	if err != nil {
-		return fmt.Errorf("Failed to chmod file %v to mode %v", dst, si.Mode())
-	}
-
-	return nil
-}
-
-// CopyDir recursively copies a directory tree, attempting to preserve permissions.
-// Source directory must exist, destination directory must *not* exist.
-// Symlinks are ignored and skipped. Credit @m4ng0squ4sh https://gist.github.com/m4ng0squ4sh/92462b38df26839a3ca324697c8cba04
-func CopyDir(src string, dst string) error {
-	src = filepath.Clean(src)
-	dst = filepath.Clean(dst)
-
-	si, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-	if !si.IsDir() {
-		return fmt.Errorf("source is not a directory")
-	}
-
-	_, err = os.Stat(dst)
-	if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	if err == nil {
-		return fmt.Errorf("destination already exists")
-	}
-
-	err = os.MkdirAll(dst, si.Mode())
-	if err != nil {
-		return err
-	}
-
-	entries, err := ioutil.ReadDir(src)
-	if err != nil {
-		return err
-	}
-
-	for _, entry := range entries {
-		srcPath := filepath.Join(src, entry.Name())
-		dstPath := filepath.Join(dst, entry.Name())
-
-		if entry.IsDir() {
-			err = CopyDir(srcPath, dstPath)
-			if err != nil {
-				return err
-			}
-		} else {
-			err = CopyFile(srcPath, dstPath)
-			if err != nil && entry.Mode()&os.ModeSymlink != 0 {
-				fmt.Printf("failed to copy symlink %s, skipping...\n", srcPath)
-				continue
-			}
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// FileExists checks a file's existence
-func FileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
 }
