+++ pkg/archive/archive_test.go
@@ -1,10 +1,13 @@
-package util_test
+package archive_test
 
 import (
+	"io/ioutil"
+	"log"
 	"os"
 	"path/filepath"
 	"testing"
 
+	"github.com/drud/ddev/pkg/archive"
 	"github.com/drud/ddev/pkg/testcommon"
 	"github.com/drud/ddev/pkg/util"
 	"github.com/stretchr/testify/assert"
@@ -19,12 +22,37 @@ var (
 	TestTarArchiveExtractDir = "wordpress-0.4.0/"
 )
 
+func TestMain(m *testing.M) {
+	// prep assets for files tests
+	testPath, err := ioutil.TempDir("", "filetest")
+	util.CheckErr(err)
+	testPath, err = filepath.EvalSymlinks(testPath)
+	util.CheckErr(err)
+	testPath = filepath.Clean(testPath)
+	TestTarArchivePath = filepath.Join(testPath, "files.tar.gz")
+
+	err = system.DownloadFile(TestTarArchivePath, TestTarArchiveURL)
+	if err != nil {
+		log.Fatalf("archive download failed: %s", err)
+	}
+
+	testRun := m.Run()
+
+	// cleanup test file
+	err = os.Remove(TestTarArchivePath)
+	if err != nil {
+		log.Fatal("failed to remove test asset: ", err)
+	}
+
+	os.Exit(testRun)
+}
+
 // TestUntar tests untar functionality, including the starting directory
 func TestUntar(t *testing.T) {
 	assert := assert.New(t)
 	exDir := testcommon.CreateTmpDir("TestUnTar1")
 
-	err := util.Untar(TestTarArchivePath, exDir, "")
+	err := archive.Untar(TestTarArchivePath, exDir, "")
 	assert.NoError(err)
 
 	// Make sure that our base extraction directory is there
@@ -40,7 +68,7 @@ func TestUntar(t *testing.T) {
 
 	// Now do the untar with an extraction root
 	exDir = testcommon.CreateTmpDir("TestUnTar2")
-	err = util.Untar(TestTarArchivePath, exDir, TestTarArchiveExtractDir)
+	err = archive.Untar(TestTarArchivePath, exDir, TestTarArchiveExtractDir)
 	assert.NoError(err)
 
 	finfo, err = os.Stat(filepath.Join(exDir, ".ddev"))
@@ -64,9 +92,9 @@ func TestUnzip(t *testing.T) {
 	assert := assert.New(t)
 	exDir := testcommon.CreateTmpDir("TestUnzip1")
 
-	zipfilePath := filepath.Join("testing", "testfile.zip")
+	zipfilePath := filepath.Join("test", "testfile.zip")
 
-	err := util.Unzip(zipfilePath, exDir, "")
+	err := archive.Unzip(zipfilePath, exDir, "")
 	assert.NoError(err)
 
 	// Make sure that our base extraction directory is there
@@ -82,7 +110,7 @@ func TestUnzip(t *testing.T) {
 
 	// Now do the unzip with an extraction root
 	exDir = testcommon.CreateTmpDir("TestUnzip2")
-	err = util.Unzip(zipfilePath, exDir, testZipExtractDir)
+	err = archive.Unzip(zipfilePath, exDir, testZipExtractDir)
 	assert.NoError(err)
 
 	// Only the dir2_file should remain
@@ -92,68 +120,4 @@ func TestUnzip(t *testing.T) {
 
 	err = os.RemoveAll(exDir)
 	assert.NoError(err)
-}
-
-// TestCopyFile tests copying a file.
-func TestCopyFile(t *testing.T) {
-	assert := assert.New(t)
-	tmpTargetDir := testcommon.CreateTmpDir("TestCopyFile")
-	tmpTargetFile := filepath.Join(tmpTargetDir, filepath.Base(TestTarArchivePath))
-
-	err := util.CopyFile(TestTarArchivePath, tmpTargetFile)
-	assert.NoError(err)
-
-	file, err := os.Stat(tmpTargetFile)
-	assert.NoError(err)
-
-	if err != nil {
-		assert.False(file.IsDir())
-	}
-	err = os.RemoveAll(tmpTargetDir)
-	assert.NoError(err)
-}
-
-// TestCopyDir tests copying a directory.
-func TestCopyDir(t *testing.T) {
-	assert := assert.New(t)
-	sourceDir := testcommon.CreateTmpDir("TestCopyDir_source")
-	targetDir := testcommon.CreateTmpDir("TestCopyDir_target")
-
-	subdir := filepath.Join(sourceDir, "some_content")
-	err := os.Mkdir(subdir, 0755)
-	assert.NoError(err)
-
-	// test source not a directory
-	err = util.CopyDir(TestTarArchivePath, sourceDir)
-	assert.Error(err)
-	assert.Contains(err.Error(), "source is not a directory")
-
-	// test destination exists
-	err = util.CopyDir(sourceDir, targetDir)
-	assert.Error(err)
-	assert.Contains(err.Error(), "destination already exists")
-	err = os.RemoveAll(subdir)
-	assert.NoError(err)
-
-	// copy a directory and validate that we find files elsewhere
-	err = os.RemoveAll(targetDir)
-	assert.NoError(err)
-
-	file, err := os.Create(filepath.Join(sourceDir, "touch1.txt"))
-	assert.NoError(err)
-	_ = file.Close()
-	file, err = os.Create(filepath.Join(sourceDir, "touch2.txt"))
-	assert.NoError(err)
-	_ = file.Close()
-
-	err = util.CopyDir(sourceDir, targetDir)
-	assert.NoError(err)
-	assert.True(util.FileExists(filepath.Join(targetDir, "touch1.txt")))
-	assert.True(util.FileExists(filepath.Join(targetDir, "touch2.txt")))
-
-	err = os.RemoveAll(sourceDir)
-	assert.NoError(err)
-	err = os.RemoveAll(targetDir)
-	assert.NoError(err)
-
 }
