+++ pkg/fileutil/files_test.go
@@ -1,0 +1,77 @@
+package fileutil_test
+
+import (
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/drud/ddev/pkg/fileutil"
+	"github.com/drud/ddev/pkg/testcommon"
+	"github.com/drud/drud-go/utils/system"
+	"github.com/stretchr/testify/assert"
+)
+
+var testFileLocation = "test/regular_file"
+
+// TestCopyDir tests copying a directory.
+func TestCopyDir(t *testing.T) {
+	assert := assert.New(t)
+	sourceDir := testcommon.CreateTmpDir("TestCopyDir_source")
+	targetDir := testcommon.CreateTmpDir("TestCopyDir_target")
+
+	subdir := filepath.Join(sourceDir, "some_content")
+	err := os.Mkdir(subdir, 0755)
+	assert.NoError(err)
+
+	// test source not a directory
+	err = fileutil.CopyDir(testFileLocation, sourceDir)
+	assert.Error(err)
+	assert.Contains(err.Error(), "source is not a directory")
+
+	// test destination exists
+	err = fileutil.CopyDir(sourceDir, targetDir)
+	assert.Error(err)
+	assert.Contains(err.Error(), "destination already exists")
+	err = os.RemoveAll(subdir)
+	assert.NoError(err)
+
+	// copy a directory and validate that we find files elsewhere
+	err = os.RemoveAll(targetDir)
+	assert.NoError(err)
+
+	file, err := os.Create(filepath.Join(sourceDir, "touch1.txt"))
+	assert.NoError(err)
+	_ = file.Close()
+	file, err = os.Create(filepath.Join(sourceDir, "touch2.txt"))
+	assert.NoError(err)
+	_ = file.Close()
+
+	err = fileutil.CopyDir(sourceDir, targetDir)
+	assert.NoError(err)
+	assert.True(system.FileExists(filepath.Join(targetDir, "touch1.txt")))
+	assert.True(system.FileExists(filepath.Join(targetDir, "touch2.txt")))
+
+	err = os.RemoveAll(sourceDir)
+	assert.NoError(err)
+	err = os.RemoveAll(targetDir)
+	assert.NoError(err)
+
+}
+
+func TestCopyFile(t *testing.T) {
+	assert := assert.New(t)
+	tmpTargetDir := testcommon.CreateTmpDir("TestCopyFile")
+	tmpTargetFile := filepath.Join(tmpTargetDir, filepath.Base(testFileLocation))
+
+	err := fileutil.CopyFile(testFileLocation, tmpTargetFile)
+	assert.NoError(err)
+
+	file, err := os.Stat(tmpTargetFile)
+	assert.NoError(err)
+
+	if err != nil {
+		assert.False(file.IsDir())
+	}
+	err = os.RemoveAll(tmpTargetDir)
+	assert.NoError(err)
+}
