+++ pkg/fileutil/test/regular_file
@@ -1,0 +1,1 @@
+This is a regular file for use in the CopyDir test.
